Unblock coordinator queue poll from LLM via _planner_lock ownership
Replace scattered concurrency flags (_queue_poll_in_flight, _preempted_tick_token,
_current_tick_token) with a single asyncio.Lock (_planner_lock) that serializes
executor ownership across the tick/self_task and coordinator queue poll paths.

Core fix: coordinator queue poll now runs as asyncio.create_task on every tick,
no longer blocked when self_task LLM is running. _planner_lock is acquired before
the HTTP /tasks/next call so no task is dequeued unless the slot is available.
If a user task arrives while self_task LLM is in progress, _thinking.discard("tick")
unblocks later ticks immediately and the LLM result is discarded.

Also fixes agent/db/reader.py ROUND(double precision) cast to ::numeric.

Known remaining concurrency issues (tracked in TODO.md):
- chat / coordinator_interrupt paths not yet under _planner_lock
- lock released after create_task, before executor._running = True (short gap)
- _check_coordinator_queue not added to session_tasks (not cancelled on reconnect)
- dequeued coordinator tasks left as running if skipped due to state conflict

+++ agent/db/reader.py
@@ -156,7 +156,7 @@ async def query_goal_completion(
                 COUNT(*) FILTER (WHERE goal_count IS NOT NULL)                          AS tasks_with_goal,
                 COUNT(*) FILTER (WHERE goal_count IS NOT NULL AND actual_count >= goal_count) AS fully_met,
                 COUNT(*) FILTER (WHERE goal_count IS NOT NULL AND actual_count < goal_count)  AS partial,
-                ROUND(AVG(actual_count::float / NULLIF(goal_count, 0)) * 100, 1)        AS avg_pct
+                ROUND((AVG(actual_count::float / NULLIF(goal_count, 0)) * 100)::numeric, 1) AS avg_pct
             FROM tasks
             WHERE created_at >= $1 AND status = 'done'{bot_filter}
             """,
